sort firmware version output.  Fix bug where firmware command was not seeting cloud URI
