+++ Source/v2/Meadow.Cli/Commands/Current/Firmware/FirmwareListCommand.cs
@@ -44,7 +44,7 @@ public class FirmwareListCommand : BaseCommand<FirmwareListCommand>
             Logger?.LogInformation($" {name}");
             var collection = manager.Firmware[name];
 
-            foreach (var package in collection)
+            foreach (var package in collection.OrderByDescending(p => p.Version))
             {
                 if (package == collection.DefaultPackage)
                 {
@@ -85,7 +85,7 @@ public class FirmwareListCommand : BaseCommand<FirmwareListCommand>
             Logger?.LogInformation($" {name}");
             var collection = manager.Firmware[name];
 
-            foreach (var package in collection)
+            foreach (var package in collection.OrderByDescending(p => p.Version))
             {
                 if (package == collection.DefaultPackage)
                 {
