+++ Source/v2/Meadow.Cloud.Client/Firmware/FirmwareClient.cs
@@ -1,16 +1,5 @@
 ﻿namespace Meadow.Cloud.Client.Firmware;
 
-public interface IFirmwareClient
-{
-    Task<IEnumerable<GetFirmwareVersionsResponse>> GetVersions(string type, CancellationToken cancellationToken = default);
-
-    Task<GetFirmwareVersionResponse?> GetVersion(string type, string version, CancellationToken cancellationToken = default);
-
-    Task<HttpResponseMessage> GetDownloadResponse(string url, CancellationToken cancellationToken = default);
-
-    Task<HttpResponseMessage> GetDownloadResponse(Uri url, CancellationToken cancellationToken = default);
-}
-
 public class FirmwareClient : MeadowCloudClientBase, IFirmwareClient
 {
     private readonly HttpClient _httpClient;
